Add get client method

+++ BankAccount.Backend/BankAccount.Domain/ViewModels/AccountViewModel.cs
@@ -5,7 +5,7 @@
         public int Id { get; set; }
         public string AccountNumber { get; set; }
         public string OpeningDate { get; set; }
-        public float Balance { get; set; }
+        public decimal Balance { get; set; }
         public string AccountType { get; set; }
         public int? CreditLimit { get; set; }
     }
